Split out functions and fixtures
